finished client and server side librarification, catching emits in the console, through rdl
